+++ lib/easy_table.dart
@@ -23,5 +23,5 @@ export 'src/row_callback_typedefs.dart';
 export 'src/cell_style.dart';
 export 'src/cell_icon.dart';
 export 'src/column_sort.dart';
-export 'src/sort_type.dart';
+export 'src/sort_order.dart';
 export 'src/last_visible_row_listener.dart';