+++ lib/src/column_sort.dart
@@ -1,9 +1,9 @@
-import 'package:easy_table/src/sort_type.dart';
+import 'package:easy_table/src/sort_order.dart';
 
 /// Describes the column sort.
 class ColumnSort {
-  ColumnSort({required this.columnIndex, required this.sortType});
+  ColumnSort({required this.columnIndex, required this.order});
 
   final int columnIndex;
-  final EasyTableSortType sortType;
+  final TableSortOrder order;
 }