+++ lib/src/sort_type.dart
@@ -1,2 +1,0 @@
-/// Sort type
-enum EasyTableSortType { ascending, descending }